requirements.txt
ccxt==4.2.35
pandas==2.2.1
python-dotenv==1.0.1
scikit-learn==1.5.1
joblib==1.4.2
flask==3.0.3
celery==5.3.6
redis==5.0.8
requests==2.32.3
pyjwt==2.9.0
